Create Model and PosynomialInequality class, implement necessary operator overloads for the simple example
Got started adding correct overloads and two new classes. Disappointingly cannot find a workaround for the Number class prototype sets, requiring a call in the file that is using gpkit.js. The correct setting code is wrapped up in setupNums. Next task is to add JSON file read in on the flask side, and conversion.