Implement VectorVariable name generation on client side
This is somewhat different to the setup with gpkit. Here we are
hacking together variable arrays, by just setting the names differ
ent.

Standard gpkit way:
- Call VectorVariable(number of vars you want,variable params...)
- Returns an array, in NomialArray format
- Each variable in the nomial array is named differently according
to index in the vectorvariable call

gpkitjs implementation:
- Call VectorVariable (JS implementation), which returns a regular
JS array of variables.
- Inside VectorVariable, names are being assigned according to idx

The key difference is that the array coming back is not specially
formatted into a NomialArray. I may implement this class later,
don't see the need right now.